Allow user defined schedulers to be supplied to thread pools
Threadmanager only needs to own a vector or pools and each
pool owns its scheduler, change api/internals to support this.

The scheduler is owned by the pool that uses it and is deleted
by the pool when it is destroyed.

threadmanager_impl is templated on the scheduler type so all
code must be in header in order for user supplied schedulers to
be instantiated.

ResourcePartitioner has a new pool creation function that takes a
callback that is invoked at pool creation time so that the user can
return a scheduler of their own to be used in the pool.

Several HPX_EXPORT additions to make code linkable with user
supplied schedulers.

The threadmanager_impl will need to be modified as it currently lies
in detail:: namespace and is required for user defined schedulers.

Adjustments to OS pool executors to fit the same framework as
the other pools etc.